Read infile and dup2 from infile fd to stdin. It can do execve to read infile

+++ pipex/process.c
@@ -1,3 +1,17 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   process.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/15 15:58:26 by jayoon            #+#    #+#             */
+/*   Updated: 2022/06/15 17:39:06 by jayoon           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "pipex.h"
+#include <unistd.h>
 
 void	execute_process(t_list *p_list, char *envp[])
 {
@@ -13,7 +27,7 @@ void	execute_process(t_list *p_list, char *envp[])
 		else
 			path = ft_add_slash_strjoin(p_list->dir_path[i], \
 									p_list->execve_argv[0]);
-		check_error(E_MALLOC, path);
+		check_error(E_MALLOC, (long long)path);
 		execve(path, p_list->execve_argv, envp);
 		ft_free_malloc(path);
 		i++;
